update file structure & add detect population from hospital_region
