+++ .gitignore
@@ -25,6 +25,7 @@ __pycache__
 
 # OS
 .DS_Store
+~$*
 
 # IDEs and editors
 /.idea
@@ -34,3 +35,14 @@ __pycache__
 *.launch
 .settings/
 *.sublime-workspace
+
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# backup file
+/data/bk
+
+# result file
+/result
