+++ handler.py
@@ -9,9 +9,48 @@ from io import BytesIO
 from dotenv import load_dotenv
 from sendgrid import SendGridAPIClient
 from sendgrid.helpers.mail import Mail
-from method.seir import gen_initial, prepare_input, seir_estimation, seir_df_to_json
+from model.seir import gen_initial, summarize_seir, get_default_params
+from model.resource_projection import get_resource_consumption
+from model_run import seir_estimation
+
 
 load_dotenv()
+
+
+def seir_df_to_json(seir_df, resource_df):
+    seir_json = seir_df.set_index('date').to_json(
+        orient='split', date_format='iso')
+    resource_json = resource_df.set_index(
+        'date').to_json(orient='split', date_format='iso')
+    return seir_json, resource_json
+
+
+def prepare_input(user_input):
+    default_params = get_default_params()
+    user_input['start_date'] = pd.to_datetime(
+        user_input.get('start_date', default_params['today']))
+    user_input['social_distancing'] = [
+        float(user_input.get('social_distancing',
+                             default_params['social_distancing_rate'])),
+        float(user_input.get('social_distancing_start',
+                             default_params['social_distance_day_start'])),
+        float(user_input.get('social_distancing_end',
+                             default_params['social_distance_day_end']))
+    ]
+    hospital_region = user_input.get('hospital_region')
+    user_input['area'] = hospital_region.split('(')[0].rstrip()
+    user_input['regional_population'] = float(hospital_region.split('(')[1].split()[0].replace(',', ''))
+    user_input['hospital_market_share'] = user_input.get(
+        'hospital_market_share', default_params['hospital_market_share'])
+    user_input['doubling_time'] = user_input.get(
+        'doubling_time', default_params['doubling_time'])
+    user_input['doubling_time'] = user_input.get(
+        'doubling_time', default_params['doubling_time'])
+    user_input['critical_cases'] = user_input.get(
+        'critical_cases', default_params['critical_cases'])
+    user_input['death'] = user_input.get(
+        'death', default_params['death'])
+    return user_input, default_params
 
 
 def supply_estimation(event, context):
@@ -19,11 +58,10 @@ def supply_estimation(event, context):
     user_input = json.loads(event['body'])
 
     # Model and prediction
-    user_input, default_params = prepare_input(user_input)
-    initial_data, params = gen_initial(default_params, user_input)
-    seir_df, resource_df = seir_estimation(
-        params, initial_data, user_input
-    )
+    model_input, default_params = prepare_input(user_input)
+    initial_data, params = gen_initial(default_params, model_input)
+    resource_consumption = get_resource_consumption()
+    seir_df, hos_load_df, resource_df = seir_estimation(params, initial_data, user_input, resource_consumption)
     seir_json, resource_json = seir_df_to_json(seir_df, resource_df)
 
     response_body = ''.join([
@@ -68,7 +106,7 @@ def supply_service(event, context):
     for index, i in enumerate(patients):
         pop_y += str(i)
         pop_x += str(index)
-        label_x += "D{}".format(str(index+1))
+        label_x += "D{}".format(str(index + 1))
         if index < (len(patients) - 1):
             pop_y += '%2C'
             pop_x += '%2C'
